Logger and package setup
The application logger is configured in the main such that any logger in other files need only set the same name to log to the same location.  package scripts have been provided to compile a jar of the application, written to support both Unix/Posix Operating Systems as well as Windows machines.  Tests directory has been configured to contain any experimental executions of the code, can be run independently as a caller to trigger the program.  Still need to setup a config file, will do after Tuesday meeting.